+++ Cargo.toml
@@ -12,9 +12,10 @@ license = "MIT OR Apache-2.0"
 [features]
 rand = ["dep:rand"]
 serde = ["dep:serde"]
+chrono = ["dep:chrono"]
 
 [dependencies]
-chrono = "0.4.39"
+chrono = { version = "0.4.39", features = ["alloc"], default-features = false, optional = true }
 rand = { version = "0.9.0", optional = true }
 serde = { version = "1.0.218", features = ["derive"], default-features = false, optional = true }
